requirements.txt
Markdown~=3.3.3
Flask~=2.2.5
Flask-Caching~=1.10.0
lru-expiring-cache~=1.1
beautifulsoup4~=4.11.0
Pygments~=2.15.0
mdx_truly_sane_lists~=1.3
pytz~=2023.3
markdown-katex~=202112.1034
watchdog~=3.0.0
markupsafe~=2.1.0
